Vector-database-Assignment/requirements.txt
python-dotenv==1.0.1

# core
PyMuPDF==1.24.14
langchain-text-splitters==0.2.4
tqdm==4.66.4

# retrieval + embeddings
numpy>=1.26,<3
faiss-cpu>=1.9.0.post1
sentence-transformers>=3.0.1

# api + ui
fastapi>=0.115
uvicorn>=0.30
streamlit>=1.41S
requests>=2.32

# generation
torch>=1.10  # Hugging Face transformers library depends on this
transformers>=4.15  # Latest Transformers for generation


sentencepiece
protobuf